pyproject.toml
[tool.poetry]
name = "ai-qna-with-your-sentiment"
version = "0.1.0"
description = ""
authors = ["bae-code <locing45@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.12"
openai = "^1.68.2"
transformers = "^4.50.1"
python-dotenv = "^1.1.0"
uvicorn = "^0.34.0"
motor = "^3.7.0"
redis = "^5.2.1"
pyjwt = "^2.10.1"
passlib = "^1.7.4"
pytest = "^8.3.5"
ruff = "^0.11.2"
pydantic-settings = "^2.8.1"
pytest-asyncio = "^0.26.0"
httpx = "^0.28.1"
torch = "^2.6.0"
openai-agents = {extras = ["viz"], version = "^0.0.9"}
anthropic = "^0.49.0"
mypy = "^1.15.0"


[tool.poetry.group.dev.dependencies]
pytest-asyncio = "^0.26.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
